+++ receiver/otlpreceiver/config.go
@@ -20,7 +20,6 @@ import (
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configgrpc"
 	"go.opentelemetry.io/collector/config/confighttp"
-	"go.opentelemetry.io/collector/config/configmap"
 )
 
 const (
@@ -55,8 +54,8 @@ func (cfg *Config) Validate() error {
 	return nil
 }
 
-// Unmarshal a config.Parser into the config struct.
-func (cfg *Config) Unmarshal(componentParser *configmap.ConfigMap) error {
+// Unmarshal a config.Map into the config struct.
+func (cfg *Config) Unmarshal(componentParser *config.Map) error {
 	if componentParser == nil || len(componentParser.AllKeys()) == 0 {
 		return fmt.Errorf("empty config for OTLP receiver")
 	}
@@ -66,7 +65,7 @@ func (cfg *Config) Unmarshal(componentParser *configmap.ConfigMap) error {
 		return err
 	}
 
-	// next manually search for protocols in the configmap.ConfigMap, if a protocol is not present it means it is disable.
+	// next manually search for protocols in the config.Map, if a protocol is not present it means it is disable.
 	protocols, err := componentParser.Sub(protocolsFieldName)
 	if err != nil {
 		return err